requirements.txt
flask==2.2.5
deepface==0.0.86
numpy==1.25.2
opencv-python-headless==4.8.1.78
Pillow==10.1.0
gunicorn==20.1.0
pytest==7.4.2
tf-keras
